adding changes

+++ src/controllers/rest/sensor.controller.ts
@@ -1,11 +1,12 @@
 import { Controller, Inject } from '@tsed/di';
 import { BodyParams, Context, PathParams, QueryParams } from '@tsed/platform-params';
-import { Get, Post, Delete, Put } from '@tsed/schema';
+import { Get, Post, Delete, Put, Patch } from '@tsed/schema';
 import { Authenticate } from '@tsed/passport';
 import { Res } from '@tsed/common';
 import { User } from '../../models/User';
 import { SensorService } from '../../services/sensor/sensor.service';
 import { UpdateSensorDto } from '../../dto/sensor/updateSensor';
+import { Role } from '../../models/Enum';
 
 @Controller('/sensor')
 export class SensorController {
@@ -19,8 +20,7 @@ export class SensorController {
             if (id.length != 24) {
                 return res.status(400).json({ sucess: false, err: 'Bad request' });
             }
-
-            const result = await this.sensorService.findById(id);
+            const result = await this.sensorService.findById({ _id: id, isDeleted: false });
 
             if (result.status === 200) {
                 return res.status(200).json({ success: true, data: result.data });
@@ -44,7 +44,10 @@ export class SensorController {
         @QueryParams('sortBy') sortBy: string,
     ) {
         try {
-            const result = await this.sensorService.find(filter, skip, take, sortBy);
+            let query = filter ? filter : {};
+            query = { ...query, ...{ isDeleted: false } };
+
+            const result = await this.sensorService.find(query, skip, take, sortBy);
             if (result.status === 200) {
                 return res.status(200).json({ success: false, data: result.data });
             } else {
@@ -71,11 +74,59 @@ export class SensorController {
         }
     }
 
+    @Put('/:id')
+    @Authenticate('jwt')
+    async update(@Context('user') user: User, @Res() res: Res, @PathParams('id') id: string, @BodyParams() body: UpdateSensorDto) {
+        try {
+            if (id.length != 24) {
+                return res.status(400).json({ sucess: false, err: 'Bad request' });
+            }
+            const sensor = await this.sensorService.findById({ _id: id, isDeleted: false });
+            if (!sensor.data) {
+                return res.status(404).json({ sucess: false, err: 'Sensor not found' });
+            }
+
+            if (user.role != Role.Admin && user._id != sensor.data.createdBy) {
+                return res.status(401).json({ sucess: false, err: 'Permission not met' });
+            }
+
+            const result = await this.sensorService.updateSensor(id, body);
+            if (result.status === 404) {
+                return res.status(404).json({ sucess: false, err: result.message });
+            } else if (result.status === 200) {
+                return res.status(200).json({ sucess: true, data: result.data });
+            } else {
+                return res.status(500).json({ sucess: false, err: result.message });
+            }
+        } catch (error) {
+            return res.status(500).json({ sucess: false, err: 'Internal server error' });
+        }
+    }
+
     @Delete('/:id')
     @Authenticate('jwt')
-    async delete(@Res() res: Res, @PathParams('id') id: string) {
+    async delete(@Context('user') user: User, @Res() res: Res, @PathParams('id') id: string, @QueryParams('force') force: boolean) {
         try {
-            const result = await this.sensorService.deleteById(id);
+            if (id.length != 24) {
+                return res.status(400).json({ sucess: false, err: 'Bad request' });
+            }
+            // Checking Permissions
+            const sensor = await this.sensorService.findById({ _id: id });
+            if (!sensor.data) {
+                return res.status(404).json({ sucess: false, err: 'Sensor not found' });
+            }
+            let role = Role.Admin;
+            if (user.role != role && user._id.toString() != sensor.data.createdBy.toString()) {
+                return res.status(401).json({ sucess: false, err: 'Requirements not met' });
+            }
+            ///
+            let result;
+            if (force == true && user.role == Role.Admin) {
+                result = await this.sensorService.forceDeleteById(id);
+            } else {
+                result = await this.sensorService.deleteById(id);
+            }
+
             if (result.status === 404) {
                 return res.status(404).json({ sucess: false, err: result.message });
             } else if (result.status === 200) {
@@ -88,21 +139,31 @@ export class SensorController {
         }
     }
 
-    @Put('/:id')
     @Authenticate('jwt')
-    async update(@Res() res: Res, @PathParams('id') id: string, @BodyParams() body: UpdateSensorDto) {
-        try {
-            if (id.length != 24) {
-                return res.status(400).json({ sucess: false, message: 'Bad request' });
-            }
+    @Patch('/restore')
+    async restore(@Context('user') user: User, @Res() res: Res, @BodyParams('ids') ids: string[]) {
+        if (!Array.isArray(ids) || ids.length === 0) {
+            return res.status(400).json({ message: 'Invalid or empty "ids" in the request body' });
+        }
 
-            const result = await this.sensorService.updateSensor(id, body);
-            if (result.status === 404) {
-                return res.status(404).json({ sucess: false, err: result.message });
-            } else if (result.status === 200) {
-                return res.status(200).json({ sucess: true, data: result.data });
+        ids.forEach((element) => {
+            if (element.length != 24) {
+                return res.status(400).json({ sucess: false, err: 'Bad request' });
+            }
+        });
+
+        const role = Role.Admin;
+        if (user.role != role) {
+            return res.status(401).json({ success: false, err: 'unauthorized' });
+        }
+
+        try {
+            const result = await this.sensorService.restore({ _id: { $in: ids }, isDeleted: true });
+
+            if (result.status === 200) {
+                return res.status(200).json({ sucess: true, err: 'Observations successfully restored' });
             } else {
-                return res.status(500).json({ sucess: false, err: result.message });
+                return res.status(result.status).json({ message: result.message });
             }
         } catch (error) {
             return res.status(500).json({ sucess: false, err: 'Internal server error' });
