+++ src/controllers/rest/sensorType.controller.ts
@@ -1,11 +1,12 @@
 import { Controller, Inject } from '@tsed/di';
 import { BodyParams, Context, PathParams, QueryParams } from '@tsed/platform-params';
-import { Get, Post, Delete, Put } from '@tsed/schema';
+import { Get, Post, Delete, Put, Patch } from '@tsed/schema';
 import { Authenticate } from '@tsed/passport';
 import { Res } from '@tsed/common';
 import { SensorTypeService } from '../../services/sensorType/sensorType.service';
 import { User } from '../../models/User';
-import { UpdateSensorTypeDto } from 'src/dto/sensorType/updateSensorDto';
+import { UpdateSensorTypeDto } from '../../dto/sensorType/updateSensorDto';
+import { Role } from '../../models/Enum';
 
 @Controller('/sensorType')
 export class SensorTypeController {
@@ -19,8 +20,7 @@ export class SensorTypeController {
             if (id.length != 24) {
                 return res.status(400).json({ sucess: false, err: 'Bad request' });
             }
-
-            const result = await this.sensorTypeService.findById(id);
+            const result = await this.sensorTypeService.findById({ _id: id, isDeleted: false });
 
             if (result.status === 200) {
                 return res.status(200).json({ success: true, data: result.data });
@@ -39,12 +39,15 @@ export class SensorTypeController {
     async find(
         @Res() res: Res,
         @QueryParams('filter') filter: string,
-        @QueryParams('skip') skip: string,
-        @QueryParams('take') take: string,
+        @QueryParams('skip') skip: number,
+        @QueryParams('take') take: number,
         @QueryParams('sortBy') sortBy: string,
     ) {
         try {
-            const result = await this.sensorTypeService.find(filter, skip, take, sortBy);
+            let query = filter ? filter : {};
+            query = { ...query, ...{ isDeleted: false } };
+
+            const result = await this.sensorTypeService.find(query, skip, take, sortBy);
             if (result.status === 200) {
                 return res.status(200).json({ success: false, data: result.data });
             } else {
@@ -94,15 +97,70 @@ export class SensorTypeController {
 
     @Delete('/:id')
     @Authenticate('jwt')
-    async delete(@Res() res: Res, @PathParams('id') id: string) {
+    async delete(@Context('user') user: User, @Res() res: Res, @PathParams('id') id: string, @QueryParams('force') force: boolean) {
         try {
-            const result = await this.sensorTypeService.deleteById(id);
+            if (id.length != 24) {
+                return res.status(400).json({ sucess: false, err: 'Bad request' });
+            }
+            // Checking Permissions
+            const sensor = await this.sensorTypeService.findById({ _id: id, isDeleted: false });
+            if (!sensor.data) {
+                return res.status(404).json({ sucess: false, err: 'SensorType not found' });
+            }
+            let role = Role.Admin;
+            if (!sensor.data.createdBy) {
+                return res.status(401).json({ sucess: false, err: 'Requirements not met' });
+            }
+
+            if (user.role != role && user._id.toString() != sensor.data.createdBy.toString()) {
+                return res.status(401).json({ sucess: false, err: 'Requirements not met' });
+            }
+            ///
+
+            let result;
+            if (force == true && user.role == Role.Admin) {
+                result = await this.sensorTypeService.forceDeleteById(id);
+            } else {
+                result = await this.sensorTypeService.deleteById(id);
+            }
+
             if (result.status === 404) {
                 return res.status(404).json({ sucess: false, err: result.message });
             } else if (result.status === 200) {
                 return res.status(200).json({ sucess: true, data: result.message });
             } else {
                 return res.status(500).json({ sucess: true, err: result.message });
+            }
+        } catch (error) {
+            return res.status(500).json({ sucess: false, err: 'Internal server error' });
+        }
+    }
+
+    @Authenticate('jwt')
+    @Patch('/restore')
+    async restore(@Context('user') user: User, @Res() res: Res, @BodyParams('ids') ids: string[]) {
+        if (!Array.isArray(ids) || ids.length === 0) {
+            return res.status(400).json({ message: 'Invalid or empty "ids" in the request body' });
+        }
+
+        ids.forEach((element) => {
+            if (element.length != 24) {
+                return res.status(400).json({ sucess: false, err: 'Bad request' });
+            }
+        });
+
+        const role = Role.Admin;
+        if (user.role != role) {
+            return res.status(401).json({ success: false, err: 'unauthorized' });
+        }
+
+        try {
+            const result = await this.sensorTypeService.restore({ _id: { $in: ids }, isDeleted: true });
+
+            if (result.status === 200) {
+                return res.status(200).json({ sucess: true, err: 'Observations successfully restored' });
+            } else {
+                return res.status(result.status).json({ message: result.message });
             }
         } catch (error) {
             return res.status(500).json({ sucess: false, err: 'Internal server error' });
